backend/requirements.txt
# ─── Core ───
pydantic<2.10
fastapi>=0.115
uvicorn>=0.30
python-multipart>=0.0.12

# ─── Database (Convex) ───
convex>=0.7

# ─── AI/ML ───
transformers>=4.44
torch>=2.1
sentence-transformers>=3.0
httpx>=0.27

# ─── Malaysian AI ASR ───
librosa>=0.10
soundfile>=0.12

# ─── OCR ───
pytesseract>=0.3.13
Pillow>=10.0

# ─── Image Processing ───
rembg>=2.0.50
onnxruntime>=1.16

# ─── Utilities ───
python-dotenv>=1.0
twilio>=9.0

# ─── OCR Fallback ───
easyocr>=1.7

# ─── Audio Processing ───
pydub>=0.25

# ─── PDF Vision (Qwen 2.5-VL via OpenRouter) ───
PyMuPDF>=1.24

# ─── Web Search Grounding (SerpAPI) ───
google-search-results>=2.4

# ─── Multi-Agent Orchestration (Snap & Sell 2.0) ───
langgraph>=0.2
langchain-core>=0.3

# ─── Video Assembly ───
moviepy>=2.0

# ─── HuggingFace Space API (LTX-2.3) ───
gradio_client>=1.0.0
